backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.1
python-dotenv==1.0.1
PyMySQL==1.1.1
cryptography==42.0.8
APScheduler==3.10.4
